Add test case with a logfile
Add a test case which passes '--logfile' to set the file to receive the
log generated. The behaviour for passing '--logfile' was changed
slightly: now every time a logfile is passed to the Single_Logger
factory, a new handler is added. Need to check if this will bring side
effects (like double log entry if two handlers are registered).