Change naming behaviour
When the --release and --name are provided together, the name provided
in --name have priority over the prefix provided in --release.
This mean if 'OLD_1_0_0' is provided in --release and 'LIBX' is provided
in --name, the used release name will be 'LIBX_1_0_0'.

In the same way, if --release and --version are provided, the version
provided in --version have priority over the version suffix provided in
--release.

Fixes #23