Remove --symbols option and make it default
This make '--symbols' default for the 'update' command.  This way the
strategy option is no longer required and will be assumed as the
same as the current '--symbols'.  This means if '--add' and '--remove'
are not provided, the provided symbols will be compared with the
existing symbols to create the set of added and removed symbols.