Added command to output version
Avoid repetition when writing to file.

Added command to get the current version.

Also added version to output. This required changes in the tests
expected output files.

Added a way to easily update the version in tests.

Updated readme and usage documentation.

Added script to get the current version to be used in bootstrap.