Fixed bugs in updates with release info
When release information were provided (using "-r"), updates which added
symbols but not removed were failing because the element used to point
to the release to be modified was obtained in a generator, which ceases
to exist in the end of the for loop.  Changing the loop to iterate over
the actual elements of the releases list solved this bug.

The strings in the result of get_info_from_release_string() were not
being converted to upper case.  This was leading to errors when
searching for releases to be updated when release info is provided.