tests: Fixed after change of logger.warning() output
The output for the warning changed, including the line where the warning
is generated.  Change the test cases to check warnings and errors using
a regular expression instead of fixed string.

Also added support for checking for error messages, which can be
different from the exception message.

Finally, added python 3.7 to the test matrix

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>